+++ datacontract/command_ci.py
@@ -62,7 +62,7 @@ def ci(
     inline_references: Annotated[
         bool,
         typer.Option(
-            help="Resolve external references (currently: authoritativeDefinitions\\[type=definition]) in the "
+            help="Resolve external references (currently: authoritativeDefinitions\\[type in {definition, semantics}]) in the "
             "contract and inline the fetched content from the configured entropy-data host."
         ),
     ] = True,