+++ datacontract/command_test.py
@@ -70,7 +70,7 @@ def test(
     inline_references: Annotated[
         bool,
         typer.Option(
-            help="Resolve external references (currently: authoritativeDefinitions\\[type=definition]) in the "
+            help="Resolve external references (currently: authoritativeDefinitions\\[type in {definition, semantics}]) in the "
             "contract and inline the fetched content from the configured entropy-data host."
         ),
     ] = True,